Pull the leaping path into the UnicornImage class
