--- unicornleap/LeapPath.swift
@@ -1,16 +1,0 @@
-import Cocoa
-
-class LeapPath {
-  let path = CGPathCreateMutable()
-
-  init(size: CGSize) {
-    let screen = NSScreen.mainScreen()!.frame
-    let origin = CGPoint(x: -size.width, y: -size.height)
-    let destination = CGPoint(x: screen.size.width + size.width, y: origin.y)
-    let midpoint = (destination.x + origin.x) / 2.0
-    let peak = size.height + 50.0
-
-    CGPathMoveToPoint(path, nil, origin.x, origin.y)
-    CGPathAddCurveToPoint(path, nil, midpoint, peak, midpoint, peak, destination.x, destination.y)
-  }
-}
